requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.1
async-timeout==4.0.2
attrs==21.4.0
bcrypt==3.2.2
black==22.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
coverage==6.3.2
ecdsa==0.17.0
fastapi==0.77.1
flake8==4.0.1
frozenlist==1.3.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
iniconfig==1.1.1
Mako==1.2.0
MarkupSafe==2.1.1
mccabe==0.6.1
multidict==6.0.2
mypy==0.961
mypy-extensions==0.4.3
packaging==21.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
pyflakes==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-asyncio==0.18.3
pytest-cov==3.0.0
pytest-mock==3.7.0
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
sqlalchemy2-stubs==0.0.2a23
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.2.0
uvicorn==0.17.6
yarl==1.7.2
